// src/app/models/interfaces-travesia.ts
export interface Travesia {
    id:                string;
    destino_id:        string;
    dificultad:        string;
    lugar:             string;
    pcia:              string;
    pcia_id:           string;
    fecha:             string;
    hora:              string;
    guia_id:           string;
    empresa:           string;
    pencuentro:        string;
    coordenadas:         string;
    desc:              string;
    ingreso:           boolean;
    detingreso:        string;
    traslado:          boolean;
    dettraslado:       string;
    desayuno:          boolean;
    rmarcha:           boolean;
    merienda:          boolean;
    detpension:        string;
    pernocte:          boolean;
    detpernocte:        string;
    botiquin:          boolean;
    detbotiquin:       string;
    csatelital:        boolean;
    cvhf:              boolean;
    detcomunicaciones: string;
    rfoto:             boolean;
    detfoto:           string;
    scarga:            boolean;
    detcarga:          string;
    imontania:         boolean;
    detindumentaria:   string;
    cequipaje:         boolean;
    detcuidado:        string;
    precio:            number;
}